+++ lib/quantum/parser.ex
@@ -12,17 +12,17 @@ defmodule Quantum.Parser do
     [r|i] = e |> String.split("/")
     [x|y] = r |> String.split("-")
     {v,_} = x |> Integer.parse
-    parse(v, y, i, min, max) |> Enum.reject(&((&1 < min) or (&1 > max)))
+    do_parse(v, y, i, min, max) |> Enum.reject(&((&1 < min) or (&1 > max)))
   end
 
-  def parse(v, [], [], _, _), do: [v]
+  defp do_parse(v, [], [], _, _), do: [v]
 
-  def parse(v, [], [i], _, _) do
+  defp do_parse(v, [], [i], _, _) do
     {x,_} = i |> Integer.parse
     [rem(v,x)]
   end
 
-  def parse(v, [y], [], min, max) do
+  defp do_parse(v, [y], [], min, max) do
     {t,_} = y |> Integer.parse
     if v < t do
       Enum.to_list(v..t)
@@ -31,9 +31,9 @@ defmodule Quantum.Parser do
     end
   end
 
-  def parse(v, y, [i], min, max) do
+  defp do_parse(v, y, [i], min, max) do
     {x,_} = i |> Integer.parse
-    parse(v, y, [], min, max) |> Enum.reject(&(rem(&1, x) != 0))
+    do_parse(v, y, [], min, max) |> Enum.reject(&(rem(&1, x) != 0))
   end
 
 end